packages/datacollector_yourquote/requirement.txt
pandas
pytube3
PyYAML